reconfigure app to use vite inplace of deprecated react configuration
components renamed with `.jsx` extensions
