--- src/components/Header.js
@@ -1,15 +1,0 @@
-import React from "react";
-import "../css/Header.css"
-
-const Header = () => {
-  return (
-    <header className="header-wrapper">
-      <p className="title">Jog Journal</p>
-      <p className="slogan">
-        Your companion for tracking every step towards your yearly goal.
-      </p>
-    </header>
-  );
-};
-
-export default Header;
